refactor(modules): give llrt_modules the responsibility for assembling modules
