+++ llrt_core/src/modules/mod.rs
@@ -3,8 +3,9 @@
 #[cfg(not(feature = "lambda"))]
 pub use llrt_modules::console;
 pub use llrt_modules::{
-    abort, assert, buffer, child_process, crypto, dns, events, exceptions, fs, http, navigator,
-    net, os, path, perf_hooks, process, stream_web, string_decoder, timers, tty, url, util, zlib,
+    abort, assert, buffer, child_process, crypto, dns, events, exceptions, fs, http,
+    module_builder, navigator, net, os, path, perf_hooks, process, stream_web, string_decoder,
+    timers, tty, url, util, zlib,
 };
 
 #[cfg(feature = "lambda")]
