+++ llrt_modules/src/lib.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+pub mod module_builder;
+
 pub use self::modules::*;
 
 mod modules {
